More robust support for local DNS, to improve VPN experience
Summary:
Add a DNS search domain config setting, allowing clients to search the
parent domain before sending the request to an external DNS server.

Implement a server-side "search domain" for the local handler, to
accomodate VPN clients that don't have the ability to configure a
client-side setting.  This allows app0.b10e.net to forward requests for
the c.peppy-breaker-161717.internal search domain to the internal
nameserver.  Resolved names are returned to the client along with an
additional CNAME answer, documenting the implicit renaming.  In theory,
this allows "ssh build0" to work over the VPN.  In practice, this only
works if the client sends a non-FQDN request, which is not a universal
behavior.

Add 'subnets' policy setting for VPNs, allowing the routing of specific
non-ring subnets.  This allows a VPN server to provide access to the
VPN's network as well as those that it manages.  In the brightgate cloud
appliance case, this allows VPN clients to reach build0 (et. al.)
without requiring that all client traffic flow over the VPN.

Factor out common functionality in the local and proxy DNS handlers
into a single handler routine.

Test Plan:
manual testing on nils-wrt and with app0.b10e.net as a VPN
endpoint for Mac and Debian clients.

Differential Revision: https://ph0.b10e.net/D1015