cl.httpd console log should have correct caller and stack
Summary:
The zap logger determines the caller location and the stack (if
requested) based on the point of the logging call.  Because of the way
`cl.httpd` handles its errors, the logging call is far from the point of
the error determination, making that information useless.

The Google Cloud Logging mechanism we use allows you to add a special
field which is used to determine the caller location and stack instead.
This change copies that mechanism, which is enabled by the `echozap`
module, so that the logs written to the console also have the better
caller locations and stacks.

Test Plan:
Run `cl.httpd` and see that errors logged on the console have useful caller
location and stack traces.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1010