minor bug fixes
Summary:
When a daemon crashes, ap.mcp issues a Fault Event and shuts down all
dependent daemons.  We should not be issuing Faults for those dependent
shut downs, as it is expected and correct behavior.

When ap.watchd sees traffic to a device that has gone offline, the stats
are associated with an empty mac address.  While we are still interested
in collecting the stats for the live half of the connection, we should
be not accumulating those stats for the offline device.

Test Plan: manual testing with nils-wrt, using both Pi and VPN clients

Differential Revision: https://ph0.b10e.net/D1013