Escrow appliance VPN private keys with Vault
Summary:
We want to be able to record an appliance's private VPN key, but do so in such a
way that the cloud services don't have access to it.  Vault allows us to define
a policy such that a service can write to a location, but not read from it, and
this commit changes `ap.rpcd` to send the key to `cl.rpcd`, which commits into
Vault.

Test Plan:
Run `cl.rpcd` on my VM, along with the `http-dev` appliance stack.  Also, deploy
onto a Model 100 and into my test MIG.

Reviewers: dp, nils

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1019