if vpn/AddKey() fails, ensure no key config state remains
Summary:
If AddKey() fails, issue a PropDelete command for the config data that
it was intended to create.  This is intended to ensure that if a
timed-out cloud command is subsequently executed on the appliance,
we will remove any created key state.  Without this follow deletion, we
would be left with an unusable wireguard key in the config tree.

Test Plan:
tested with svc1 and dogfood, stopping ap.rpcd to simulate a
network outage.  Verified that the cloud-created key was deleted when
ap.rpcd was restarted.  Verified that a locally-created key with the
same mac address was not deleted.

Differential Revision: https://ph0.b10e.net/D1021