When deleting a config subtree, uncache all descendent nodes as well
Summary:
When we remove a subtree from the config tree, any cached values for
leaf nodes or descendent subtrees were left in the cache.  If the
deleted subtree was subsequently rebuilt, Get operations against that
subtree could return the stale data from the cache.

Test Plan:
manual testing on my local system.  Added a new test to the
configd test suite specifically to check for this failure.

Differential Revision: https://ph0.b10e.net/D1032