Make `cl-release status` emit human-readable appliance names
Summary:
The output of `cl-release status` is frustrating because the appliances are
named only by UUID, leaving the user to do a series of lookups with `cl-reg` if
they want to understand whose appliance each one is.  This changes the default
output to emit organization, site, and appliance names, putting the original
UUID output behind the `-n` flag.

Test Plan: Run resulting `cl-release` on `svc1`, plus unit tests.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1033