Add support to cl.httpd to pull GCP access tokens from Vault
Summary:
These tokens replace the use of credentials files that sit in the cloud
image.  The GCP APIs have support for ephemeral tokens, and are able to
pull from a `TokenSource` interface, for which we provide an
implementation pulling the tokens from Vault.

Test Plan:
Run this in my load-balanced cloud setup, without the environment variable
`GOOGLE_APPLICATION_CREDENTIALS` set and no credentials file available, and
confirm that the service starts up and loads avatars properly.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1039