store node mode in the config tree
Summary:
When ap.rpcd starts up, it will store the mode of the node it is running
on in the config tree.  This information can be used to guide the UI in
deciding which options to present to the admin.

Test Plan: manual testing in nils-wrt and app0.b10e.net

Differential Revision: https://ph0.b10e.net/D1044