UX fixes for `cl-reg app new` and Vault
Summary:
As we transition to using Vault, it may be convenient to both write the new
appliance cloud secret to Vault and to a local file.  This preserves the
default, but if an output directory is specified with `-d`, then `cl-reg` will
attempt to write the file there.  Note that the default when `-d` is not
specified is no longer the current directory, but empty, signifying the secret
should not be written to a file at all.

This change also tests the Vault token's capabilities against the path to which
the secret will be written, and if the check fails or the capabilities are
insufficient, `cl-reg` will abort before creating the secret and registering the
appliance in the database.

Also included is a fix for an empty vault path when escrow fails.

Test Plan: Manual testing.

Reviewers: dp, stephen

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1046