Add the rdpscan package to the list of artifacts
Summary:
The release code is somewhat non-robust in that it doesn't discover all the
artifacts for a release or the order in which they should be installed.  When we
added `rdpscan` to rXS, we forgot to update rPS to reflect that.

Test Plan:
Run `cl-release` on `svc1` to re-create Beta 1 RC1, and see that the `mt7623`
arch has `bg-rdpscan_....ipk` in its release descriptor.

Reviewers: dp, stephen

Reviewed By: stephen

Differential Revision: https://ph0.b10e.net/D1051