ap-factory harden, ramdisk fixes
Summary:
This changeset fixes a number of annoyances when `ap-factory` is invoked from a
ramdisk boot, rather than a normal boot from the eMMC drive.  It also
introduces the `harden` subcommand, which is used at the end of a factory
installation process.

Test Plan: Built, tested to install three MT7623 systems.

Reviewers: dp, duvall

Reviewed By: duvall

Differential Revision: https://ph0.b10e.net/D1058