Implement fuzzy org matching, use it for cl-release status and cl-reg site list
Summary:
We have had the ability to match sites by name, fuzzily.  This adds the ability
to do so for organizations as well, and adds the ability for `cl-reg` to list
sites and appliances by organization.

It also tweaks the way matches are delivered, and provides for some tweaking
from the command-line, for experimentation.

Test Plan: Unit testing and spot-checking against the existing database.

Reviewers: dp, stephen

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1063