Provide the cloud with more information about failed upgrades
Summary:
This change allows the cloud to have a clearer picture of what an appliance is
running, and how its last upgrade fared.

We now record the commit hashes an appliance reports, what stage of upgrade the
appliance is known to have either passed (successfully or otherwise), as well as
the error text of whatever caused the failure and the URL of the install output
log, if it got that far.

All this information is now shown in `cl-release status`, though the error and
log URL are only displayed with `-v`.

This also includes some small cleanups, and adds tests to `cl.eventd`.

Test Plan:
Run unit tests.  Manual testing running both cloud and appliance out of the
workspace.

Reviewers: dp, stephen

Reviewed By: dp, stephen

Differential Revision: https://ph0.b10e.net/D1061