Handle renamed/removed .go files better
Summary:
When you remove or rename a `.go` source file, either directly with `git mv` or
`git rm`, or when switching between branches that have different files, or
pulling new changes from upstream, you'll see an error like this:

    make: *** No rule to make target 'golang/src/bg/ap.networkd/hostapd.go', needed by '_tools.amd64/godeps.mk'.  Stop.

The remedy is to remove `_tools.amd64/godeps.mk`, but why should you do that by
hand, like a sucker?  This change has `make` do it automatically: if it detects
that any `.go` file is missing, it'll remove `godeps.mk` and rebuild it, before
continuing on to whatever you asked it to do.

Test Plan:
Switch between branches where a file gets removed.  Demonstrate the error
without the fix, demonstrate that it continues with it.

Reviewers: dp, stephen, nils

Reviewed By: stephen, nils

Differential Revision: https://ph0.b10e.net/D1067