Clean up some cosmetic issues after D1061
Summary:
The "Current Release UUID" column in `cl-release status` shouldn't be
empty when the UUID is the nil UUID.  Previously, this would have shown
the all-zero string, but this changes it to "-" for slightly easier
viewing.  The corresponding name goes back to being "Unknown/Mixed"
instead of "-".

`cl.eventd` was also needlessly complaining about the nil UUID release.

The schema file was added at the wrong number, so renumber it.

Test Plan:
Add a few new unit tests, but also check the output of `cl-release` status by
hand.

Reviewers: dp, stephen

Reviewed By: stephen

Differential Revision: https://ph0.b10e.net/D1071