Let cl.rpcd pull credentials from Vault
Summary:
`cl.rpcd` should pull its credentials from Vault, just like `cl.httpd` does.
This includes database login and password, as well as access tokens to perform
the GCP tasks it has.

The piece most likely to fail is the `vaulttoken` retry mechanism, which has
been introduced to deal with the service account being replaced, which can
happen when the roleset it uses has its resource bindings changed.  This should
be a rare event, but without automatic retry could lead a daemon to present a
useless access token for the rest of the token's lifetime, which could be up to
an hour.

Test Plan:
Run appliance and cloud from a workspace, and run the cloud bits in
`dd-dev-cloud1`.

Reviewers: dp

Reviewed By: dp

Differential Revision: https://ph0.b10e.net/D1075