Simplify the ring selection logic
Summary:
This removes the non-intuitive "allowable ring upgrade" logic.  Each
client has an optional 'home' ring.  When it connects to a VAP that
hosts that ring, the client will be assigned to it.  If it connects to a
different VAP, or has no designated home ring, it will be assigned to
the default ring for that VAP.  This will allow any client to
successfully connect to any VAP, while not losing track of the ring the
admin really wants it to use.

The default ring for the EAP VAP is now 'standard' rather than
'unenrolled'.

These new defaults and behaviors are less secure than the old model, but
should be significantly less confusing to users and admins.

Test Plan: manual testing with wired and wireless Pi and Mac clients

Differential Revision: https://ph0.b10e.net/D1077