Fix vpn-related firewall rule construction
Summary:
When constructing the list of vpn-related firewall rules, the result
should be a concatenated list of server-side rules and client-side
rules.  Instead, the client-side rules were replacing the server-side
rules.  This change fixes that bug.

Test Plan:
manual on nils-wrt. Verified that the board was able to act
as a VPN client and server simultaneously

Differential Revision: https://ph0.b10e.net/D1079