Prevent unintended migration to "unenrolled"
Summary:
Clients are assigned to rings by ap.configd when it receives a NetEntity
event.  Those events may arrive from ap.serviced (in response to a DHCP
request) or ap.wifid (in response to new wireless connection).

There is a brief window in which a new wireless client is connected to
the "unenrolled" ring, before ap.wifid receives the ring assignment from
ap.configd and migrates the device to the correct VLAN for that ring.
In that window, ap.serviced may receive a DHCP request on the unenrolled
VLAN, send a NetEntity event to ap.configd indicating that it found the
client on the unenrolled ring, which will cause ap.configd to reassign
the client to that ring permanently.

This change prevents ap.configd from responding to the DHCP-triggered
NetEntity event sent while a wireless client is still connected to the
unenrolled VLAN.

Test Plan:
manual testing with a mac client on a clean, default configuration.
Repeatedly switched the client between all three SSIDs, verifying that it
was assigned to the correct default for that SSID each time.  Set the
client's home ring to "core", and verified that it correctly returned to
that ring each time it connected to the EAP VAP.

Differential Revision: https://ph0.b10e.net/D1080