Added a sanity check to vpntool
Summary:
Getting a VPN fully configured involves setting multiple parameters, and
ensuring that different settings and files be kept in sync.  This complexity has
led to roll-out failures at multiple sites, which have taken substantial time to
diagnose.  This addition to ap-vpntool and cl-vpntool functions as an automated
checklist, searching for common configuration errors.

In addition to checking the existing config tree, this also adds a new
server_key setting to each user key configuration.  This new setting
will allow us to identify keys that have become stale or invalid because
of a change in the private server key on the appliance.

Test Plan: manual testing with different failure scenarios on nils-wrt.
Verified that the new cl-vpntool would validate VPN configs on old
appliances, and that it would still succesfully manipulate VPN keys
without the new property being available.

Differential Revision: https://ph0.b10e.net/D1087