Improve security for cloud and appliance cookies
Summary:
Enable 4 security attributes for the session cookies we issue.  This
helps to tamp down a set of possible security problems associated with
cookies:
  - Secure: mark the cookie so that it is only ever transmitted over HTTPS
  - HttpOnly: mark the cookie as inaccessible to JS Document.Cookie APIs
  - SameSite=Strict: mark the cookie to be used in first-party context only

Additionally we rename the cookies with a special prefix recognized by
modern browsers which reinforces the above, and shorten up the lifespan
to a week.

Test Plan:
- Ran in simulated cloud and simulated appliance for regression testing
- Tested against simulated XSS attack
- Simulated sending cookie without HttpOnly/Secure set, saw browser reject it

Reviewers: stephen, duvall

Reviewed By: stephen, duvall

Differential Revision: https://ph0.b10e.net/D1088